requirements.txt
streamlit
langchain
langchain-community
markdown
sentence-transformers
unstructured
python-dotenv
google-generativeai
